requirements.txt
groq
psycopg2-binary
fastapi
uvicorn
